# PERF: DataFrame construction