```
In [1]: from asv_bench.benchmarks.frame_ctor import *

In [2]: cls = FromArrays

In [3]: self = cls()

In [4]: self.setup()

In [5]: %timeit self.time_frame_from_arrays_int()
6.21 ms ± 184 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)  # <- master
3.27 ms ± 182 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)  # <- PR

```