@jreback 2021-07-22
cool i would actually add a release note about this as this is a significant perf improvement. does this hold more generally with other construction asv's?

also doe you think its the removing asserts or the changes in not using is_sparse that are the big wins here? (guessing its the asserts)