@jbrockmendel 2021-07-22
> also doe you think its the removing asserts or the changes in not using is_sparse that are the big wins here? (guessing its the asserts)

I forget the exact numbers, but remember being surprised how big a difference the is_sparse change made