@rhshadrach 2021-07-24
I'm seeing failure for the groupby.apply doctest on master. The condition that leads to raising an error is

    if newb.shape != self.shape:

Looks like shape is being cached here - might be related?