@jbrockmendel 2021-07-24
good catch, that caching should be reverted (and ideally a non-doctest test implemented)